+++ packages/mui-base/src/NumberInputUnstyled/NumberInputUnstyled.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import PropTypes from 'prop-types';
 import { OverridableComponent } from '@mui/types';
 import classes from './numberInputUnstyledClasses';
-import useNumberInput from './useNumberInput';
+import useNumberInput from '../useNumberInput';
 import {
   NumberInputUnstyledOwnerState,
   NumberInputUnstyledProps,
