+++ packages/mui-base/src/useNumberInput/index.ts
@@ -1,0 +1,3 @@
+export { default } from './useNumberInput';
+
+export * from './useNumberInput.types';
